+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - feat: change `NesTabView` to hide close icons when no handler is informed
  - feat: add `NesContainerTheme` to theming for improved customization of `NesContainer`s
  - feat: add `NesButtonTheme`.borderColor for improved customization of `NesButton`s
+ - feat: change `NesTab`to use the same border color as `NesContainer`s for design consistency. 
 
 # 0.7.0